fix: 경로 수정, topbar 삭제
